# python-client/swagger_client/models/account.py
# coding: utf-8

"""
    Buycycle

    DevOps homework  # noqa: E501

    OpenAPI spec version: 1.0.0
    
    Generated by: https://github.com/swagger-api/swagger-codegen.git
"""


import pprint
import re  # noqa: F401

import six


class Account(object):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    """
    Attributes:
      swagger_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    swagger_types = {
        'deals': 'list[str]',
        'name': 'str',
        'persons': 'list[str]',
        'transfers': 'list[str]'
    }

    attribute_map = {
        'deals': 'deals',
        'name': 'name',
        'persons': 'persons',
        'transfers': 'transfers'
    }

    def __init__(self, deals=None, name=None, persons=None, transfers=None):  # noqa: E501
        """Account - a model defined in Swagger"""  # noqa: E501

        self._deals = None
        self._name = None
        self._persons = None
        self._transfers = None
        self.discriminator = None

        if deals is not None:
            self.deals = deals
        self.name = name
        if persons is not None:
            self.persons = persons
        if transfers is not None:
            self.transfers = transfers

    @property
    def deals(self):
        """Gets the deals of this Account.  # noqa: E501


        :return: The deals of this Account.  # noqa: E501
        :rtype: list[str]
        """
        return self._deals

    @deals.setter
    def deals(self, deals):
        """Sets the deals of this Account.


        :param deals: The deals of this Account.  # noqa: E501
        :type: list[str]
        """

        self._deals = deals

    @property
    def name(self):
        """Gets the name of this Account.  # noqa: E501


        :return: The name of this Account.  # noqa: E501
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this Account.


        :param name: The name of this Account.  # noqa: E501
        :type: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def persons(self):
        """Gets the persons of this Account.  # noqa: E501


        :return: The persons of this Account.  # noqa: E501
        :rtype: list[str]
        """
        return self._persons

    @persons.setter
    def persons(self, persons):
        """Sets the persons of this Account.


        :param persons: The persons of this Account.  # noqa: E501
        :type: list[str]
        """

        self._persons = persons

    @property
    def transfers(self):
        """Gets the transfers of this Account.  # noqa: E501


        :return: The transfers of this Account.  # noqa: E501
        :rtype: list[str]
        """
        return self._transfers

    @transfers.setter
    def transfers(self, transfers):
        """Sets the transfers of this Account.


        :param transfers: The transfers of this Account.  # noqa: E501
        :type: list[str]
        """

        self._transfers = transfers

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.swagger_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value
        if issubclass(Account, dict):
            for key, value in self.items():
                result[key] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, Account):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other